Use the `with` statement on the connection object
Before, v.0.0.3, the connection was reinitialized everytime a new
auth request was sent to the module. But after that version, things
changed and a single connection was used.

This caused #2. It was because the transactions weren't commited
before querying again. Which stops the auth provider from getting
new data. So, the context manager is also used on the connection
everytime a query is sent to ensure the transactions are committed.

So, hopefully this commit fixes #2